Text search fixes, numeric field min/max/step, default controller improvements

+++ lib/active_element/default_controller.rb
@@ -9,6 +9,8 @@ module ActiveElement
     end
 
     def index
+      return render_forbidden(:listable) unless configured?(:listable)
+
       controller.render 'active_element/default_views/index',
                         locals: {
                           collection: collection,
@@ -17,16 +19,22 @@ module ActiveElement
     end
 
     def show
+      return render_forbidden(:viewable) unless configured?(:viewable)
+
       controller.render 'active_element/default_views/show', locals: { record: record }
     end
 
     def new
+      return render_forbidden(:editable) unless configured?(:editable)
+
       controller.render 'active_element/default_views/new', locals: { record: model.new, namespace: namespace }
     end
 
-    def create # rubocop:disable Metrics/AbcSize
+    def create # rubocop:disable Metrics/AbcSize, Metrics/MethodLength
+      return render_forbidden(:editable) unless configured?(:editable)
+
       new_record = model.new(default_record_params.params)
-      # Ensure associations are applied:
+      # XXX: Ensure associations are applied - there must be a better way.
       if new_record.save && new_record.reload.update(default_record_params.params)
         controller.flash.notice = "#{new_record.model_name.to_s.titleize} created successfully."
         controller.redirect_to record_path(new_record, :show).path
@@ -34,13 +42,19 @@ module ActiveElement
         controller.flash.now.alert = "Failed to create #{model.name.to_s.titleize}."
         controller.render 'active_element/default_views/new', locals: { record: new_record, namespace: namespace }
       end
+    rescue ActiveRecord::RangeError => e
+      render_range_error(error: e, action: :new)
     end
 
     def edit
+      return render_forbidden(:editable) unless configured?(:editable)
+
       controller.render 'active_element/default_views/edit', locals: { record: record, namespace: namespace }
     end
 
     def update # rubocop:disable Metrics/AbcSize
+      return render_forbidden(:editable) unless configured?(:editable)
+
       if record.update(default_record_params.params)
         controller.flash.notice = "#{record.model_name.to_s.titleize} updated successfully."
         controller.redirect_to record_path(record, :show).path
@@ -48,9 +62,13 @@ module ActiveElement
         controller.flash.now.alert = "Failed to update #{model.name.to_s.titleize}."
         controller.render 'active_element/default_views/edit', locals: { record: record, namespace: namespace }
       end
+    rescue ActiveRecord::RangeError => e
+      render_range_error(error: e, action: :edit)
     end
 
     def destroy
+      return render_forbidden(:deletable) unless configured?(:deletable)
+
       record.destroy
       controller.flash.notice = "Deleted #{record.model_name.to_s.titleize}."
       controller.redirect_to record_path(model, :index).path
@@ -60,12 +78,26 @@ module ActiveElement
 
     attr_reader :controller
 
+    def render_forbidden(type)
+      controller.render 'active_element/default_views/forbidden', locals: { type: type }
+    end
+
+    def configured?(type)
+      return state.deletable? if type == :deletable
+
+      state.public_send("#{type}_fields").present?
+    end
+
+    def state
+      @state ||= controller.active_element.state
+    end
+
     def default_record_params
       @default_record_params ||= ActiveElement::DefaultRecordParams.new(controller: controller, model: model)
     end
 
     def default_text_search
-      @default_text_search ||= ActiveElement::DefaultTextSearch.new(controller: controller, model: model)
+      @default_text_search ||= ActiveElement::DefaultSearch.new(controller: controller, model: model)
     end
 
     def record_path(record, type = nil)
@@ -88,6 +120,18 @@ module ActiveElement
       return model.all unless default_text_search.text_search?
 
       model.left_outer_joins(default_text_search.search_relations).where(*default_text_search.text_search)
+    end
+
+    def render_range_error(error:, action:)
+      controller.flash.now.alert = formatted_error(error)
+      controller.render "active_element/default_views/#{action}", locals: { record: record, namespace: namespace }
+    end
+
+    def formatted_error(error)
+      return error.cause.message.split("\n").join(', ') if error.try(:cause)&.try(:message).present?
+      return error.message if error.try(:message).present?
+
+      I18n.t('active_element.unexpected_error')
     end
   end
 end
